chore: working android project Refactored the android project so that it is a lib project used in a sample app. That way the lib project can be published as an aar to be used elsewhere
