+++ Android/app/build.gradle
@@ -1,40 +1,27 @@
 apply plugin: "com.android.application"
-apply plugin: "kotlin-android"
-apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 33
     defaultConfig {
         applicationId "ademar.androidioscppexample"
-        minSdkVersion 14
-        targetSdkVersion 27
+        minSdkVersion 21
+        targetSdkVersion 33
         versionCode 1
         versionName "1.0"
-        externalNativeBuild {
-            cmake {
-                cppFlags "-std=c++14"
-            }
-        }
     }
     buildTypes {
         release {
             minifyEnabled true
-            useProguard true
             shrinkResources false
             proguardFiles getDefaultProguardFile("proguard-android.txt")
-        }
-    }
-    externalNativeBuild {
-        cmake {
-            path "CMakeLists.txt"
         }
     }
 }
 
 dependencies {
-    implementation "com.android.support:appcompat-v7:27.0.1"
-    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation project(path: ':cppTest')
 }
 repositories {
     mavenCentral()
