--- Android/app/src/main/java/ademar/androidioscppexample/KotlinActivity.kt
@@ -1,16 +1,0 @@
-package ademar.androidioscppexample
-
-import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import kotlinx.android.synthetic.main.activity_kotlin.*
-
-class KotlinActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_kotlin)
-
-        text.text = CoreWrapper().concatenateMyStringWithCppString("Kotlin")
-    }
-
-}
