+++ Android/app/src/main/res/layout/activity_main.xml
@@ -14,14 +14,4 @@
         android:textSize="26sp"
         tools:text="Hello World!"/>
 
-    <Button
-        android:id="@+id/button"
-        style="@style/Base.Widget.AppCompat.Button.Colored"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|center"
-        android:layout_marginBottom="16dp"
-        android:text="@string/app_click_kotlin"
-        android:textColor="@android:color/white"/>
-
 </merge>
